requirements.txt
azure-cognitiveservices-speech
python-dotenv
requests
moviepy